+++ Vol1A/SVD/testDriver.py
@@ -1,11 +1,11 @@
-# solutions.py
+# testDriver.py
 from scipy import linalg as la
 import numpy as np
 from matplotlib import pyplot as plt
 
 # Problem 1
 def truncated_svd(A,r=None,tol=10**-6):
-    """Computes the truncated SVD of A. If r is None or equals the number 
+    """Computes the truncated SVD of A. If r is None or equals the number
         of nonzero singular values, it is the compact SVD.
     Parameters:
         A: the matrix
@@ -44,13 +44,13 @@ def visualize_svd(A):
     """Plot each transformation associated with the SVD of A."""
     S = np.load("circle.npz")["circle"]
     vec = np.load("circle.npz")["unit_vectors"]
-    
+
     U,s,Vh = la.svd(A)
-    
+
     VhS = Vh.dot(S)
     SigVhS = np.diag(s).dot(VhS)
     USigVhS = U.dot(SigVhS)
-    
+
     Vhvec = Vh.dot(vec)
     SigVhvec = np.diag(s).dot(Vhvec)
     USigVhvec = U.dot(SigVhvec)
@@ -78,11 +78,11 @@ def visualize_svd(A):
 # Problem 3
 def svd_approx(A, k):
     """Returns best rank k approximation to A with respect to the induced 2-norm.
-    
+
     Inputs:
     A - np.ndarray of size mxn
-    k - rank 
-    
+    k - rank
+
     Return:
     Ahat - the best rank k approximation
     """
@@ -94,16 +94,16 @@ def svd_approx(A, k):
         print "WARNING: Given parameters do not result in compressed data."
     Ahat = U[:,:k].dot(S).dot(Vt[:k,:])
     return Ahat
-    
+
 # Problem 4
 def lowest_rank_approx(A,e):
-    """Returns the lowest rank approximation of A with error less than e 
+    """Returns the lowest rank approximation of A with error less than e
     with respect to the induced 2-norm.
-    
+
     Inputs:
     A - np.ndarray of size mxn
     e - error
-    
+
     Return:
     Ahat - the lowest rank approximation of A with error less than e.
     """
@@ -111,14 +111,14 @@ def lowest_rank_approx(A,e):
 
     k = np.where(s<e)[0][0]
     print k
-    
+
     return svd_approx(A,k)
-    
+
 # Problem 5
 def compress_image(filename,k):
     """Plot the original image found at 'filename' and the rank k approximation
     of the image found at 'filename.'
-    
+
     filename - jpg image file path
     k - rank
     """
@@ -126,68 +126,68 @@ def compress_image(filename,k):
     red = orig_img[:,:,0]
     green = orig_img[:,:,1]
     blue = orig_img[:,:,2]
-    
+
     img = np.zeros(orig_img.shape)
     img[:,:,0] = svd_approx(red,k)
     img[:,:,1] = svd_approx(green,k)
     img[:,:,2] = svd_approx(blue,k)
-    
+
     img = np.round(img)/255.
     orig_img = np.round(orig_img)/255.
     img[img<0] = 0.
     img[img>1] = 1.
-    
+
     plt.subplot(2,1,1)
-    plt.title("Original Image")    
+    plt.title("Original Image")
     plt.imshow(orig_img)
     plt.subplot(2,1,2)
     plt.title("Rank " + str(k) + " Approximation")
     plt.imshow(img)
     plt.show()
-    
+
 def grayImage(filename):
 
     img_color = plt.imread(filename)
     return (img_color[:,:,0]+img_color[:,:,1]+img_color[:,:,2])/3.0
-    
-    
+
+
 def test(student_module, late=False):
-    
+
     def grade(p):
         """Manually grade a problem worth 'p' points"""
         part = -1
         while part > p or part < 0:
             part = int(input("\nScore out of " + str(p) + ": "))
         if part == p: return p,""
-        else: 
+        else:
             m = "\n" + raw_input("\nEnter extra feedback:")
             return part,m
-            
+
     def arrayTest(A,B,f):
         if np.allclose(A,B,rtol=.1, atol=.1):
             return 1, ""
         else:
             f = f + "\nCorrect answer:\n" + str(A) + "\nStudent answer:\n" + str(B) + "\n\n"
             return 0, f
-            
+
     s = student_module
-    
+
     score = 0
     total = 50
     feedback = ""
-        
+
     try: #Problem 1: 10 points
         feedback += "\n\nProblem 1 (10 points):"
-        points = 0   
+        points = 0
         errMessage = "\nIncorrect SVD for A = \n"
         m = 4
         n = 3
         A = np.array([1, 0, 3, 1, 2, 0, 5, 0, -1, 0, 3, 5]).reshape((m,n))
         r = 2
         U, S, Vh = truncated_svd(A,r)
-        
+
         U2,S2,Vh2 = s.truncated_svd(A,r)
-        
+
         #Check matrix shapes
         if S2.size != r or U2.shape != (m,r) or Vh2.shape != (r,n):
             print U2.shape, S2.shape, Vh2.shape
@@ -196,22 +196,22 @@ def test(student_module, late=False):
             if U2.shape == (m,r) and Vh2.shape == (r,n) and S2.shape == (r,r):
                 feedback += "\nHint: return just the diagonal of S (as specified in the docstring)"
                 points += 3
-            
+
         else:
             S = np.diag(S)
             S2 = np.diag(S2)
             p,f = arrayTest(U.dot(S.dot(Vh)), U2.dot(S2.dot(Vh2)), errMessage+str(A))
             points += p*5
             feedback += f
-            
+
         m = 10
         n = 5
         A = np.random.randint(0,300,(m,n))
         r = 5
         U, S, Vh = truncated_svd(A,r)
-        
+
         U2,S2,Vh2 = s.truncated_svd(A,r)
-        
+
         #Check matrix shapes
         if S2.size != r or U2.shape != (m,r) or Vh2.shape != (r,n):
             print U2.shape, S2.shape, Vh2.shape
@@ -226,13 +226,13 @@ def test(student_module, late=False):
             p,f = arrayTest(U.dot(S.dot(Vh)), U2.dot(S2.dot(Vh2)), errMessage+str(A))
             points += p*5
             feedback += f
-            
+
         score += points
-        feedback += "\nScore += " + str(points)        
-    
+        feedback += "\nScore += " + str(points)
+
     except Exception as e:
         feedback += "\nError: " + e.message
-        
+
     try: #Problem 2: 10 points
         feedback += "\n\nProblem 2 (10 points):"
         points = 0
@@ -240,13 +240,13 @@ def test(student_module, late=False):
         p, f = grade(10)
         points = p
         feedback += f
-        
+
         score += points
         feedback += "\nScore += " + str(points)
-            
+
     except Exception as e:
         feedback += "\nError: " + e.message
-        
+
     try: #Problem 3: 10 points
         feedback += "\n\nProblem 3 (10 points):"
         points = 0
@@ -256,34 +256,34 @@ def test(student_module, late=False):
         b = 3
         C = np.random.rand(18,10)
         c = 4
-        
+
         A1 = svd_approx(A,a)
         A2 = s.svd_approx(A,a)
         if A2 is not None:
             p,f = arrayTest(A1,A2,"\nWrong approximation for k=29, A 30x30")
             points += p*4
             feedback += f
-        
+
         B1 = svd_approx(B,b)
         B2 = s.svd_approx(B,b)
         if B2 is not None:
             p,f = arrayTest(B1,B2,"\nWrong approximation for k=3, A 6x8")
             points += p*3
             feedback += f
-        
+
         C1 = svd_approx(C,c)
         C2 = s.svd_approx(C,c)
         if C2 is not None:
             p,f = arrayTest(C1,C2,"\nWrong approximation for k=3, A 6x8")
             points += p*3
             feedback += f
-        
+
         score += points
         feedback += "\nScore += " + str(points)
-            
+
     except Exception as e:
         feedback += "\nError: " + e.message
-        
+
     try: #Problem 4: 10 points
         points = 0
         feedback += "\n\nProblem 4 (10 points):"
@@ -295,7 +295,7 @@ def test(student_module, late=False):
             p,f = arrayTest(H1,H2,"\nWrong approximation for the Hubble image, e=3.5")
             points += p*5
             feedback += f
-        
+
         e2 = 10.0
         H1 = lowest_rank_approx(hubble,e2)
         H2 = s.lowest_rank_approx(hubble,e2)
@@ -303,13 +303,13 @@ def test(student_module, late=False):
             p,f = arrayTest(H1,H2,"\nWrong approximation for the Hubble image, e=10.0")
             points += p*5
             feedback += f
-        
+
         score += points
         feedback += "\nScore += " + str(points)
-        
+
     except Exception as e:
         feedback += "\nError: " + e.message
-        
+
     try: #Problem 5: 10 points
         '''
         Spots on image = -2
@@ -320,24 +320,24 @@ def test(student_module, late=False):
         p,f = grade(5)
         points += p
         feedback += f
-        
+
         s.compress_image("portland.jpg",40)
         p,f = grade(5)
         points += p
         feedback += f
-        
+
         score += points
         feedback += "\nScore += " + str(points)
-            
+
     except Exception as e:
         feedback += "\nError: " + e.message
-        
-        
+
+
     if late:    # Late submission penalty
         feedback += "\n\nHalf credit for late submission."
         feedback += "\nRaw score: " + str(score) + "/" + str(total)
         score *= .5
-    
+
     # Report final score.
     feedback += "\n\nTotal score: " + str(score) + "/" + str(total)
     percentage = (100.0 * score) / total
@@ -345,11 +345,11 @@ def test(student_module, late=False):
     if   percentage >=  98.0: feedback += "\n\nExcellent!"
     elif percentage >=  90.0: feedback += "\n\nGreat job!"
     return score, feedback
-        
-        
-        
-        
-        
-        
-        
-        
+
+
+
+
+
+
+
+
