+++ Vol2B/scipyoptimize/blackbox_function.py
@@ -1,12 +1,12 @@
 # blackbox_function.py
-"""Volume 2: Optimization Packages I (scipy.optimize). Auxiliary File."""
+"""Volume 2B: Optimization Packages I (scipy.optimize). Auxiliary File."""
 
 import numpy as np
 from scipy import linalg as la
 
 def blackbox(y_free):
-    """
-    Finds the length of a curve approximated piece-wise by a set of points.
+    """Find the length of a curve approximated piece-wise by a set of points.
+
     Accepts:
         y_free (1xn ndarray): the non-endpoint y-values of the curve.
 
@@ -17,14 +17,12 @@ def blackbox(y_free):
     m = len(y_free) + 2 # Number points: free variables, origin, and endpoint.
     a, b = 40, 30       # Coordinates of endpoint.
 
-     # Generate the evenly-spaced x-values of the curve.
+    # Generate the evenly-spaced x-values of the curve.
     x = np.linspace(0,a,m)
 
-     # Pad the free variables with the fixed endpoint values, 0 and b.
+    # Pad the free variables with the fixed endpoint values, 0 and b.
     y = np.hstack((0,y_free, b))
 
-     # Calculate and return the line integral of the approximated curve.
-    partial_norms = []
-    for i,item in enumerate(y[:-1]):
-        partial_norms.append(la.norm(np.array([x[i+1]-x[i],y[i+1] - item])))
-    return np.sum(partial_norms)
+    # Calculate and return the line integral of the approximated curve.
+    return sum([la.norm(np.array([x[i+1]-x[i],y[i+1]]))
+                                            for i,item in enumerate(y[:-1])])
